changed username to integer field
I am not able to create reservations with some issue, meaning that some of the types must not be working together. Checking the db table, it appears that reservation takes a user id num as a second param, not a name. Trying to use just a num doesn't seem to work so I'm wondering if it isn't just a string of the num from the string field. So I am trying it as an Integer Field now.